+++ src/logic/interaction/UserInteractionInterface.h
@@ -10,7 +10,7 @@ class UserInteractionInterface {
 public:
     UserInteractionInterface(UserInteraction* reference);
 
-    void add_tower(const TowerType& type, const Position<double>& position);
+    void add_tower(const TowerType& type, const Position<int>& position);
 private:
     UserInteraction* user_interaction_;
 };
