+++ src/logic/tower/Tower.h
@@ -11,7 +11,7 @@ class WorldState;
 
 class Tower {
 public:
-    static Tower create_tower(WorldState* world_ref, const TowerType& ref, const Position<double>& position);
+    static Tower create_tower(WorldState* world_ref, const TowerType& ref, const Position<int>& position);
     ~Tower();
     // No copies or assignments are allowed (only moves)
     Tower(const Tower&) = delete;
@@ -22,7 +22,7 @@ public:
     const int& get_cost() const;
     const std::vector<TowerRotation>& get_requested_rotations() const;
     const std::vector<double>& get_requested_shoots() const;
-    const Position<double>& get_position() const;
+    const Position<int>& get_position() const;
     const double& get_rotational_speed() const;
     double& get_angle();
 
@@ -32,14 +32,14 @@ public:
     void clear_requests();
 protected:
     Tower(WorldState*, const int& damage, const int& radar_load_time, const int& cost, const int& range,
-        const double& rotational_speed, const Position<double>& pos);
+        const double& rotational_speed, const Position<int>& pos);
 private:
     static unsigned int instance_counter;
 
     const unsigned int id_;
     const int damage_, radar_load_time_, cost_, range_;
     const double rotational_speed_;
-    const Position<double> pos_;
+    const Position<int> pos_;
 
     double angle_;
     std::vector<double> requested_shoots_;
